requirements.txt
PyQt5>=5.15.7
aiohttp>=3.8.0
aiofiles>=23.0.0
PyYAML>=6.0
cryptography>=41.0.0
Pillow>=10.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-docx>=0.8.11
jinja2>=3.1.0
schedule>=1.2.0
